Move seed service to Infrastructure
