+++ src/Application/Services/ApplicationService.cs
@@ -1,5 +1,4 @@
-﻿using Application.Infrastructure;
-using Application.Interfaces;
+﻿using Application.Interfaces;
 using Application.Models;
 using Application.Workflows;
 using Microsoft.Extensions.AI;
