+++ src/Infrastructure/Extensions/InfrastructureExtensions.cs
@@ -1,5 +1,6 @@
 ﻿using Application.Interfaces;
 using Infrastructure.Repository;
+using Infrastructure.Services;
 using Microsoft.Extensions.DependencyInjection;
 
 namespace Infrastructure.Extensions;
@@ -14,6 +15,9 @@ public static class InfrastructureExtensions
         services.AddScoped<IArtifactRepository, ArtifactRepository>();
 
         services.AddScoped<IWorkflowRepository, WorkflowRepository>();
+
+        services.AddHostedService<AzureStorageSeedService>();
+
 
         return services;
     }
